// mrpSystem/src/app/shared/models/mainView.model.ts
export interface IMainView {
    TempSeqId: string;
    SeqId: number;
    JobNo: string;
    QuotationNo: string;
    RevisionNo: number;
    ProposalNo: string;
    MedicalType: string;
    PolicyNo: string;
    LoanAmount: number;
    Interest: number;
    Term: number;
    FullTermInMonths: number;
    GracePeriod: number;
    CompanyBufferId: number;
    CurrentAwplr: number;
    AdditionToAwplr: number;
    TermOfFixedInterest: number;
    BankId: number;
    BranchId: number;
    CurrencyId: number;
    InterestRateType: string;
    HnbaBranchCode: string;
    BrokerCode: number;
    ChannelCode: number;
    IsReInsurance: number;
    LoanTypeId: number;
    ReInsCompanyId: number;
    ExchangeRate: number;
    DateOfCommence: string;
    DateOfProposal: string;
    Premium: number;
    PremiumWithPolicyFee: number;
    Status: string;
    UserId: string;
    ProposalSendingMethod: string;
    RegisterDate: string;
    IsValidated: number;
    IsVIP: number;
    BankName: string;
    BankBranchName: string;
    HnbaBranch: string;
    LoanType: string;
}